+++ src/support/slack/commands.js
@@ -63,6 +63,10 @@ import addOaeStageDomain from './commands/add-oae-stage-domain.js';
 import togglePathSuggestions from './commands/toggle-path-suggestions.js';
 import getPathSuggestionsStatus from './commands/get-path-suggestions-status.js';
 import brandClaims from './commands/toggle-brand-claims.js';
+import triggerImpactMeasurement from './commands/trigger-impact-measurement.js';
+import checkImpactMeasurement from './commands/check-impact-measurement.js';
+import getExperiment from './commands/get-experiment.js';
+import enrichBrandClaims from './commands/enrich-brand-claims.js';
 
 /**
  * Returns all commands.
@@ -124,4 +128,8 @@ export default (context) => [
   togglePathSuggestions(context),
   getPathSuggestionsStatus(context),
   brandClaims(context),
+  triggerImpactMeasurement(context),
+  checkImpactMeasurement(context),
+  getExperiment(context),
+  enrichBrandClaims(context),
 ];